fix(resilience): address PR #376 review feedback
- Replace Task.Delay timing with FakeTimeProvider for deterministic tests
  (eliminates CI flakiness from timing variance)
- Split ICircuitBreaker.cs into separate files to reduce hot-file coupling:
  - CircuitState.cs
  - CircuitBreakerOptions.cs
  - CircuitBreakerStatistics.cs
  - CircuitBreakerEventArgs.cs
  - CircuitBreakerOpenException.cs
  - IBreakerRegistry.cs
  - ICircuitBreaker.cs (interface only)
- Add TimeProvider parameter to BrainarrCircuitBreakerAdapter for test control
- Add GetStatistics_Returns_NonNull_Core_Fields_After_Operations test
  (verifies stats contract wasn't silently degraded)
