+++ Brainarr.Plugin/Services/Resilience/CircuitBreakerOpenException.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace NzbDrone.Core.ImportLists.Brainarr.Services.Resilience
+{
+    /// <summary>
+    /// Exception thrown when a circuit breaker is open and blocking requests.
+    /// </summary>
+    public class CircuitBreakerOpenException : Exception
+    {
+        public CircuitBreakerOpenException(string message) : base(message) { }
+    }
+}
